requirements.txt
streamlit
pillow
numpy
opencv-python
streamlit-image-select
base64